% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.methdf5.sheet.R
\name{read.methdf5.sheet}
\alias{read.methdf5.sheet}
\title{process as sample sheet directly into an RGChannelSet backed by HDF5}
\usage{
read.methdf5.sheet(sheet, filepath = "my_h5_se", ...)
}
\arguments{
\item{sheet}{a sample sheet, as from read.metharray.sheet}

\item{filepath}{a directory in which to put the HDF5 data ("my_h5_se")}

\item{...}{other arguments, passed to minfi:::read.metharray2}
}
\value{
an HDF5-backed RGChannelSet
}
\description{
For processing lists of IDATs, see read.methdf5
}
\examples{

# cheezy 450k example
if (require("minfiData")) {
  baseDir <- system.file("extdata", package = "minfiData")
  sheet <- read.metharray.sheet(baseDir)
  inCore <- read.metharray(basenames=sheet$Basename)
  outOfCore <- read.methdf5.sheet(sheet)
  stopifnot(identical(dim(inCore), dim(outOfCore)))
  verifyRGsets(ram=inCore, hdf5=outOfCore)
}

}
\seealso{
read.methdf5
}
