% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabixToGr.R
\name{tabixedToGr}
\alias{tabixedToGr}
\title{Scan Tabix'ed files into a merged GRanges\link{List} (for RaggedExperiments).}
\usage{
tabixedToGr(...)
}
\arguments{
\item{...}{TABIXed files from which to generate a joint GenomicRanges}
}
\value{
a GRanges
}
\description{
Modified from a prototype originally by Wanding Zhou (github.com/zwdzwd).
}
\examples{
\dontrun{
  gr <- tabixedToGr(list.files(patt=".hg19.bed.gz$"))
}

}
