% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabixToGr.R
\name{tabixFileListToGr}
\alias{tabixFileListToGr}
\title{Scan a TabixFileList into a joint GRanges by union()'ing each tabixed file's}
\usage{
tabixFileListToGr(tfl, verbose = TRUE)
}
\arguments{
\item{tfl}{a TabixFileList}

\item{verbose}{squawk? (TRUE)}
}
\value{
a GRanges
}
\description{
Scan a TabixFileList into a joint GRanges by union()'ing each tabixed file's
}
\examples{
\dontrun{
  tfl <- TabixFileList(lapply(list.files(patt=".hg19.bed.gz$"), TabixFile))
  gr <- tabixFileListToGr(tfl) 
}

}
