% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verifyRGsets.R
\name{verifyRGsets}
\alias{verifyRGsets}
\title{Verify that an RGChannelSet backed by HDF5 files matches an in-core twin.}
\usage{
verifyRGsets(ram, hdf5, rows = NULL)
}
\arguments{
\item{ram}{the in-core RGChannelSet}

\item{hdf5}{the hdf5-backed RGChannelSet}

\item{rows}{how many rows to test? (default: ncol(ram))}
}
\value{
a logical of length one
}
\description{
Verify that an RGChannelSet backed by HDF5 files matches an in-core twin.
}
\seealso{
read.methdf5
}
